# BasicCRUD/docs/swagger.yaml
basePath: /
definitions:
  models.Car:
    properties:
      brand:
        type: string
      id:
        type: integer
      model:
        type: string
      name:
        type: string
      price:
        type: number
      year:
        type: integer
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is an API spec for car inventory management
  title: Car Inventory
  version: "1.0"
paths:
  /cars:
    post:
      consumes:
      - application/json
      description: Adds a new car to inventory
      parameters:
      - description: Car to add
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.Car'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new car
      tags:
      - cars
  /cars/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a car from inventory
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a car
      tags:
      - cars
    get:
      consumes:
      - application/json
      description: Get a car from inventory
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a car
      tags:
      - cars
    put:
      consumes:
      - application/json
      description: Update a car in inventory
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: string
      - description: Car to update
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Car'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a car
      tags:
      - cars
swagger: "2.0"